+++ neo/client_test.go
@@ -8,20 +8,27 @@ import (
 )
 
 func TestClient(t *testing.T) {
-	nodeURI := selectTestNode()
+	nodes := []string{
+		"http://seed5.neo.org:10332",
+		"http://seed4.neo.org:10332",
+		"http://seed3.neo.org:10332",
+		"http://seed2.neo.org:10332",
+		"http://seed1.neo.org:10332",
+	}
 
 	t.Run("NewClient()", func(t *testing.T) {
-		client := neo.NewClient(nodeURI)
+		client := neo.NewClient(nodes[0])
 
-		assert.Equal(t, nodeURI, client.NodeURI)
+		assert.Equal(t, nodes[0], client.Node)
 		assert.IsType(t, neo.Client{}, client)
 	})
 
 	t.Run(".GetBestBlockHash()", func(t *testing.T) {
 		t.Run("HappyCase", func(t *testing.T) {
-			client := neo.NewClient(nodeURI)
-			blockHash, err := client.GetBestBlockHash()
+			client, err := neo.NewClientUsingMultipleNodes(nodes)
+			assert.NoError(t, err)
 
+			blockHash, err := client.GetBestBlockHash()
 			assert.NoError(t, err)
 			assert.NotEmpty(t, blockHash)
 		})
@@ -29,7 +36,8 @@ func TestClient(t *testing.T) {
 
 	t.Run(".GetBlockByHash()", func(t *testing.T) {
 		t.Run("HappyCase", func(t *testing.T) {
-			client := neo.NewClient(nodeURI)
+			client, err := neo.NewClientUsingMultipleNodes(nodes)
+			assert.NoError(t, err)
 
 			for _, testBlock := range testBlocks {
 				t.Run(testBlock.id, func(t *testing.T) {
@@ -47,7 +55,8 @@ func TestClient(t *testing.T) {
 
 	t.Run(".GetBlockByIndex()", func(t *testing.T) {
 		t.Run("HappyCase", func(t *testing.T) {
-			client := neo.NewClient(nodeURI)
+			client, err := neo.NewClientUsingMultipleNodes(nodes)
+			assert.NoError(t, err)
 
 			for _, testBlock := range testBlocks {
 				t.Run(testBlock.id, func(t *testing.T) {
@@ -65,9 +74,10 @@ func TestClient(t *testing.T) {
 
 	t.Run(".GetBlockCount()", func(t *testing.T) {
 		t.Run("HappyCase", func(t *testing.T) {
-			client := neo.NewClient(nodeURI)
-			blockCount, err := client.GetBlockCount()
+			client, err := neo.NewClientUsingMultipleNodes(nodes)
+			assert.NoError(t, err)
 
+			blockCount, err := client.GetBlockCount()
 			assert.NoError(t, err)
 			assert.NotEmpty(t, blockCount)
 		})
@@ -75,7 +85,8 @@ func TestClient(t *testing.T) {
 
 	t.Run(".GetBlockHash()", func(t *testing.T) {
 		t.Run("HappyCase", func(t *testing.T) {
-			client := neo.NewClient(nodeURI)
+			client, err := neo.NewClientUsingMultipleNodes(nodes)
+			assert.NoError(t, err)
 
 			for _, testBlockHash := range testBlockHashes {
 				t.Run(testBlockHash.id, func(t *testing.T) {
@@ -91,9 +102,10 @@ func TestClient(t *testing.T) {
 
 	t.Run(".GetConnectionCount()", func(t *testing.T) {
 		t.Run("HappyCase", func(t *testing.T) {
-			client := neo.NewClient(nodeURI)
-			blockCount, err := client.GetConnectionCount()
+			client, err := neo.NewClientUsingMultipleNodes(nodes)
+			assert.NoError(t, err)
 
+			blockCount, err := client.GetConnectionCount()
 			assert.NoError(t, err)
 			assert.NotEmpty(t, blockCount)
 		})
@@ -101,7 +113,8 @@ func TestClient(t *testing.T) {
 
 	t.Run(".GetStorage()", func(t *testing.T) {
 		t.Run("HappyCase", func(t *testing.T) {
-			client := neo.NewClient(nodeURI)
+			client, err := neo.NewClientUsingMultipleNodes(nodes)
+			assert.NoError(t, err)
 
 			storage, err := client.GetStorage(
 				"0xecc6b20d3ccac1ee9ef109af5a7cdb85706b1df9",
@@ -115,7 +128,8 @@ func TestClient(t *testing.T) {
 
 	t.Run(".GetTransaction()", func(t *testing.T) {
 		t.Run("HappyCase", func(t *testing.T) {
-			client := neo.NewClient(nodeURI)
+			client, err := neo.NewClientUsingMultipleNodes(nodes)
+			assert.NoError(t, err)
 
 			for _, testTransaction := range testTransactions {
 				t.Run(testTransaction.id, func(t *testing.T) {
@@ -131,7 +145,8 @@ func TestClient(t *testing.T) {
 
 	t.Run(".GetTransactionOutput()", func(t *testing.T) {
 		t.Run("HappyCase", func(t *testing.T) {
-			client := neo.NewClient(nodeURI)
+			client, err := neo.NewClientUsingMultipleNodes(nodes)
+			assert.NoError(t, err)
 
 			for _, testTransactionOutput := range testTransactionOutputs {
 				t.Run(testTransactionOutput.id, func(t *testing.T) {
@@ -150,18 +165,20 @@ func TestClient(t *testing.T) {
 
 	t.Run(".GetUnconfirmedTransactions()", func(t *testing.T) {
 		t.Run("HappyCase", func(t *testing.T) {
-			client := neo.NewClient(nodeURI)
-			_, err := client.GetUnconfirmedTransactions()
+			client, err := neo.NewClientUsingMultipleNodes(nodes)
+			assert.NoError(t, err)
 
+			_, err = client.GetUnconfirmedTransactions()
 			assert.NoError(t, err)
 		})
 	})
 
 	t.Run(".Ping()", func(t *testing.T) {
 		t.Run("HappyCase", func(t *testing.T) {
-			client := neo.NewClient(nodeURI)
-			ok := client.Ping()
+			client, err := neo.NewClientUsingMultipleNodes(nodes)
+			assert.NoError(t, err)
 
+			ok := client.Ping()
 			assert.True(t, ok)
 		})
 
@@ -169,8 +186,8 @@ func TestClient(t *testing.T) {
 			for _, testPing := range testPings {
 				t.Run(testPing.description, func(t *testing.T) {
 					client := neo.NewClient(testPing.uri)
-					ok := client.Ping()
 
+					ok := client.Ping()
 					assert.False(t, ok)
 				})
 			}
@@ -178,7 +195,8 @@ func TestClient(t *testing.T) {
 
 		t.Run(".ValidateAddress()", func(t *testing.T) {
 			t.Run("HappyCase", func(t *testing.T) {
-				client := neo.NewClient(nodeURI)
+				client, err := neo.NewClientUsingMultipleNodes(nodes)
+				assert.NoError(t, err)
 
 				for _, testAccount := range testAccounts {
 					t.Run(testAccount.publicAddress, func(t *testing.T) {
@@ -191,9 +209,10 @@ func TestClient(t *testing.T) {
 			})
 
 			t.Run("SadCase", func(t *testing.T) {
-				client := neo.NewClient(nodeURI)
-				isValid, err := client.ValidateAddress("wake-up-neo")
+				client, err := neo.NewClientUsingMultipleNodes(nodes)
+				assert.NoError(t, err)
 
+				isValid, err := client.ValidateAddress("wake-up-neo")
 				assert.NoError(t, err)
 				assert.False(t, isValid)
 			})