+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this
 project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## v1.9.0 - 2018-01-22
+
+### Added
+
+- New function to select the best node from a array of node URIs.
+
+```golang
+client, err := neo.NewClientUsingMultipleNodes(
+  []string{
+    "http://seed1.neo.org:10332",
+    "http://seed2.neo.org:10332",
+    "http://seed3.neo.org:10332",
+  },
+)
+
+err = client.SelectBestNode()
+```
+
 ## v1.8.0 - 2017-10-27
 
 ### Added